Decouple MainFragment from SensonrEventListener, create SensorHandler
